+++ packages/sass/index.js
@@ -4,10 +4,13 @@
  * @see https://github.com/jtangelder/sass-loader
  */
 
+const _ = require('lodash')
+
 module.exports = sass
 
 /**
  * @param {object}   [options]                  See https://github.com/sass/node-sass#options
+ * @param {bool}     [options.minimize]         Enable minification.
  * @param {string[]} [options.includePaths]
  * @param {bool}     [options.indentedSyntax]
  * @param {string}   [options.outputStyle]
@@ -15,6 +18,7 @@ module.exports = sass
  * @return {Function}
  */
 function sass (options = {}) {
+  const sassOptions = _.omit(options, 'minimize')
   return (context, util) => util.addLoader(
     Object.assign({
       test: /\.(sass|scss)$/,
@@ -23,12 +27,13 @@ function sass (options = {}) {
         {
           loader: 'css-loader',
           options: {
-            sourceMap: Boolean(options.sourceMap)
+            sourceMap: Boolean(options.sourceMap),
+            minimize: options.minimize
           }
         },
         {
           loader: 'sass-loader',
-          options
+          options: sassOptions
         }
       ]
     }, context.match)